プレイヤーのモーション、モデル切り替え完了

+++ Source/Graphics/GaussianBlurShader.cpp
@@ -202,7 +202,7 @@ void GaussianBlurShader::CalcGaussianFilter(CBFilter& cbFilter, const GaussianFi
 	// ‹ô”‚Ìê‡‚ÍŠï”‚É’¼‚·
 	if (kernelSize % 2 == 0) kernelSize++;
 	kernelSize = max(1, min(MaxkernelSize - 1, kernelSize));
-	cbFilter.kernelSize = kernelSize;
+	cbFilter.kernelSize = (float)kernelSize;
 	cbFilter.texcel.x = 1.0f / gaussianFilterData.textureSize.x;
 	cbFilter.texcel.y = 1.0f / gaussianFilterData.textureSize.y;
